==> notebooks/modeling/05-base-strategy/02-split-pairs.ipynb <==
%run ./../../utils/commonImports.py
%matplotlib inline

pairs = pd.read_csv('pairs.tsv', header=None)[0].values
np.random.shuffle(pairs)

pd.Series(pairs[:30]).sort_values().to_csv('pairs_train.tsv', index=False)

pd.Series(pairs[30:]).sort_values().to_csv('pairs_test.tsv', index=False)
