==> notebooks/modeling/05-base-strategy/03-learn-strategy-params.ipynb <==
%run ../../utils/commonImports.py
%run ../../utils/tradingImports.py
%matplotlib inline

from base import supres

# Load Data

# dropbox_dir = 'D:\\Dropbox\\My work\\krypl-project'
dropbox_dir = '/Users/david.vesely/Dropbox/My work/krypl-project'
def read_data(pair):
    db = os.path.join(dropbox_dir,'sqlite', 'ploniex-chart-data', f'{pair}.db')
    data = load_trading_data(db, 'chart_data', from_date='2015-01-01', period='30min')
    data['ohlc4'] = (data['open'] + data['close'] + data['high'] + data['low']) / 4
    return data

# Core functions

def get_supports(data, window_size):
    supports, _ = supres(data['ohlc4'].values, window_size)
    supports = pd.DataFrame(supports, columns=['index', 'support'])
    support_vals = supports['support'].values
    size = support_vals.shape[0]
    _return = (support_vals[1:] / support_vals[:size-1]) - 1
    _return = np.insert(_return, 0, 0)
    supports['return1'] = _return
    return supports

def labels(supports, ohlc, target, stop_loss):
    cols = ['low', 'high', 'close']
    low_c, high_c, close_c = 0, 1, 2
    np_ohlc = ohlc[cols].values

    def get_label(start_i, price):
        for j in range(start_i, np_ohlc.shape[0]):
            r_low = (np_ohlc[j, low_c] / price) - 1
            r_high = (np_ohlc[j, high_c] / price) - 1
            if r_low <= stop_loss:
                return 'stop'
            if r_high >= target:
                return 'target'
        return 'unknown'

    

    labels = []
    for i in range(supports.shape[0]):
        row_s = supports.iloc[i, :]
        start = int(row_s['index'] + 1)
        price = np_ohlc[start, close_c]
        label = get_label(start+1, price)
        labels.append(label)
    
    return pd.Series(labels)

def target_hits(supports, _return):
    labels = supports[supports['return1'] <= _return].reset_index()['label']
    distb = labels.value_counts().reset_index()
    distb.columns = ['label', 'cnt']
    distb['rel'] = distb['cnt'] / distb['cnt'].sum()
    target = distb[distb['label'] == "target"]
    if target.shape[0] == 0:
        return 0., 0.
    
    return target['cnt'].iloc[0], target['rel'].iloc[0]

# Parallelize

from multiprocessing import Pool
import time
from tqdm import *

def run_parallel(f, args, n_process=10):
    results = []
    with Pool(processes=n_process) as p:
        all_args = list(args)
        with tqdm_notebook(enumerate(p.imap_unordered(f, all_args)), leave=False, total=len(all_args)) as pbar:
            for i, x in pbar:
                results.append(x)
                pbar.update()
                
    return results

def dict_from_list(l):
    d = {}
    for row in l:
        d_inner = d
        for k in row[:-2]:
            if k not in d_inner.keys():
                d_inner[k] = {}
            d_inner = d_inner[k]
        d_inner[row[-2]] = row[-1]
    return d

# Experiment setup

train_coins = pd.read_csv('pairs_train.tsv', header=None)[0].tolist()
window_sizes = [4, 6, 8, 10, 12, 24, 30, 48]
targets = list(np.linspace(0, 0.13, 14))[1:]
stop_losses = list(-np.linspace(0, 0.1, 11))[1:]
returns = list(-np.linspace(0, 0.1, 11))[1:]

loops = len(train_coins) * len(window_sizes) * len(targets) * len(stop_losses) * len(returns)
loops

# Calc Supports

data_dict = {pair: read_data(pair) for pair in train_coins}

from itertools import product

def supports_f(args):
    pair, window_size = args
    return pair, window_size, get_supports(data_dict[pair], window_size)

args = product(train_coins, window_sizes)
results = run_parallel(supports_f, args)
supports_dict = dict_from_list(results)
save_model(supports_dict, 'supports.pkl')

# Calc support labels

supports_dict = load_model('supports.pkl')

from itertools import product

def labels_f(args):
    pair, window_size, target, stop_loss = args
    data = data_dict[pair]
    supports = supports_dict[pair][window_size]
    return pair, window_size, target, stop_loss, labels(supports, data, target, stop_loss)

args = product(train_coins, window_sizes, targets, stop_losses)
results = run_parallel(labels_f, args)
labels_dict = dict_from_list(results)
save_model(labels_dict, 'labels.pkl')

# Calc Hits

supports_dict = load_model('supports.pkl')
labels_dict = load_model('labels.pkl')

from itertools import product

def target_hits_f(args):
    pair, window_size, target, stop_loss, _return = args
    data = data_dict[pair]
    supports = supports_dict[pair][window_size].copy()
    supports['label'] = labels_dict[pair][window_size][target][stop_loss]
    hits, hits_rel = target_hits(supports, _return)
    
    return pair, window_size, target, stop_loss, _return, hits, hits_rel

args = product(train_coins, window_sizes, targets, stop_losses, returns)
results = run_parallel(target_hits_f, args)
columns = ['pair', 'window_size', 'target', 'stop_loss', '_return', 'target_hits', 'target_hits_rel']
result_df = pd.DataFrame(results, columns=columns)

rpath = os.path.join(dropbox_dir,'result', 'base-strategy-learn.tsv')
write_tsv(result_df, rpath)
