==> notebooks/modeling/05-base-strategy/07-test-data.ipynb <==
%run ../../utils/commonImports.py
%run ../../utils/tradingImports.py
%matplotlib inline

from base import supres

# Load Data

# dropbox_dir = 'D:\\Dropbox\\My work\\krypl-project'
dropbox_dir = '/Users/david.vesely/Dropbox/My work/krypl-project'
def read_data(pair):
    db = os.path.join(dropbox_dir,'sqlite', 'ploniex-chart-data', f'{pair}.db')
    data = load_trading_data(db, 'chart_data', from_date='2018-01-01', period='30min')
    data['ohlc4'] = (data['open'] + data['close'] + data['high'] + data['low']) / 4
    return data

all_pairs = pd.read_csv('pairs.tsv', header=None)[0].tolist()
data_dict = {pair: read_data(pair) for pair in all_pairs}

# Missing and duplicate candles

def dates_l(start, end, delta):
    size = int((end - start) / delta) + 1
    deltas = pd.Series([delta]*size)
    dates = deltas.cumsum() + start - delta
    return dates.apply(lambda x: x.strftime("%Y-%m-%d %H:%M:%S"))

data['date'].iloc[0]

from datetime import timedelta  

for pair in all_pairs:
    data = data_dict[pair]
    min_date = str_time_to_datetime(data['date'].min())
    max_date = str_time_to_datetime(data['date'].max())
    
    dates = set(dates_l(min_date, max_date, timedelta(minutes=30)))
    data_dates = set(data['date'])
    
    missing = list(dates - data_dates)
    date_counts = data['date'].value_counts()
    duplicates = list(date_counts[date_counts > 1].index)
    
    data_dates = data['date'].tolist()
    is_sorted = all(data_dates[i] <= data_dates[i + 1] for i in range(len(data)-1))
            
    print(f"---- {pair} ----")
    print("missing: ", ','.join(missing))
    print("duplicates: ", ','.join(duplicates))
    print("sorted: ", is_sorted)
    print("")

# Pairs Volume

volume_stats = pd.DataFrame()
for pair in all_pairs:
    data = data_dict[pair]
    pair_stats = data[['volume']].describe().transpose()
    pair_stats['sum'] = data['volume'].sum()
    pair_stats['pair'] = pair
    volume_stats = volume_stats.append(pair_stats)

cols = list(volume_stats.columns)
cols = cols[-1:] + cols[:-1]
volume_stats = reset_index_hard(volume_stats).sort_values('mean', ascending=False)[cols]

volume_stats

# Find best params

cols = ['window_size', 'target', 'stop_loss', '_return']
aggregated = results.groupby(cols)\
    .mean()\
    .reset_index()

### Based on total profit

DAY = 48
aggregated.query('numberOfTrades > 200')\
    .query(f'maxTimeToClose < {3*DAY}')\
    .sort_values(['totalProfit', 'winPercentage'], ascending=False)\
    .head(10)

### Based on win percentage

DAY = 48
selected = aggregated.query('numberOfTrades > 200')\
    .query(f'maxTimeToClose < {3*DAY}')\
    .query('winPercentage > 80')\
    .sort_values(['totalProfit'], ascending=False)\
    .head(10)

selected_first = selected.iloc[0, :]
selected

cols = ['window_size', 'target', 'stop_loss', '_return']
selected_results = results
for c in cols:
    selected_results = selected_results.query(f'{c} == {selected_first[c]}')

selected_results
