==> notebooks/modeling/05-base-strategy/10-conclusion.ipynb <==
# Conclusion

Base strategy je urcite way to go, ale je potreba:
    
* vybrat jen top coiny, ktere se stabilne obchoduji
* MultiStrategy musi byt komplexnejsi a ne jen N BaseStrategy najednou
    

Dale jsou to moznosti ke zlepseni:
    
* Pohyblivy stoploss
* Ucit i multistrategy:
    * max pocet obchodu
    * s jakym procentem obchodovat
